Cleanup code / file-per-class.
