+++ src/ExtendedXmlSerializer/ContentModel/Converters/OptimizedConverterAlteration.cs
@@ -21,19 +21,15 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System;
-using System.Reflection;
 using ExtendedXmlSerialization.Core.Sources;
-using ExtendedXmlSerialization.Core.Specifications;
 
 namespace ExtendedXmlSerialization.ContentModel.Converters
 {
-	class CachedConverter : Converter<object>
+	class OptimizedConverterAlteration : IAlteration<IConverter>
 	{
-		public CachedConverter(IConverter converter) : this(converter, converter.Parse, converter.Format) {}
+		public static OptimizedConverterAlteration Default { get; } = new OptimizedConverterAlteration();
+		OptimizedConverterAlteration() {}
 
-		public CachedConverter(ISpecification<TypeInfo> specification, Func<string, object> deserialize,
-		                       Func<object, string> serialize)
-			: base(specification, new Cache<string, object>(deserialize).Get, new Cache<object, string>(serialize).Get) {}
+		public IConverter Get(IConverter parameter) => new CachedConverter(parameter);
 	}
 }
