+++ src/ExtendedXmlSerializer/TypeModel/ITypePartitions.cs
@@ -21,13 +21,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System;
 using System.Reflection;
 using ExtendedXmlSerialization.Core.Sources;
 
 namespace ExtendedXmlSerialization.TypeModel
 {
-	public delegate Func<string, TypeInfo> Partition(string parameter);
-
 	public interface ITypePartitions : IParameterizedSource<Assembly, Partition> {}
 }
